// users-service/src/modules/error-mapping/endpoint-name.enum.ts
export enum EndpointName {
  //LOS
  INITIATE_WORKFLOW = 'INITIATE_WORKFLOW',
  INTERACTION_TARGET = 'INTERACTION_TARGET',
  INTERACTION_OUTCOME = 'INTERACTION_OUTCOME',

  //LMS
  DASHBOARD = 'DASHBOARD',

  //School Aggregator
  GET_STUDENT_DETAILS = 'GET_STUDENT_DETAILS',
  WHITELIST = 'WHITELIST',

  //Refinitiv
  SCREENING_REQUEST = 'SCREENING_REQUEST',

  //MTN
  OPT_IN = 'OPT_IN',
}